requirements.txt
Flask==2.1.1
python-dotenv==0.20.0
pandas==1.4.2
requests-cache==0.9.4